requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-cors-headers==4.3.1
six==1.16.0
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
dj-rest-auth==2.2.4
celery[redis]==5.4.0
redis==5.0.8
django-celery-results==2.5.1
django-celery-beat==2.7.0